insurance_agents/requirements_azure.txt
# Azure AI Foundry Requirements for Insurance Agents
# Install with: pip install -r requirements_azure.txt

# Azure AI Foundry Core
azure-ai-projects>=1.0.0
azure-identity>=1.15.0
azure-core>=1.29.0

# Azure AI Services
azure-ai-ml>=1.12.0
azure-cognitiveservices-language-textanalytics>=5.3.0

# Azure Storage (for documents and logs)
azure-storage-blob>=12.19.0
azure-storage-file-datalake>=12.14.0

# Azure Key Vault (for secrets management)
azure-keyvault-secrets>=4.7.0

# HTTP and API clients
aiohttp>=3.9.0
httpx>=0.25.0
requests>=2.31.0

# FastAPI for registry dashboard
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data processing
pandas>=2.1.0
numpy>=1.25.0
pydantic>=2.5.0

# Insurance-specific libraries
python-dotenv>=1.0.0
jsonschema>=4.20.0

# Logging and monitoring
structlog>=23.2.0
prometheus-client>=0.19.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.10.0
flake8>=6.1.0

# Optional: Jupyter for development
jupyter>=1.0.0
ipykernel>=6.26.0
